Added our logo to the top of the page
To be honest, I'm not 100% happy with how that looks, I might have a
play arround with it to see if I can move it to the left more, I feel
like it pushes the menu down the page too far. But like all things with
design it's very subjective.